Disabling open in view pattern

+++ src/main/java/torimia/superheroes/award/service/AwardServiceImpl.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import torimia.superheroes.award.AwardMapper;
 import torimia.superheroes.award.model.dto.AwardDto;
 import torimia.superheroes.award.model.entity.Award;
@@ -11,23 +12,26 @@ import torimia.superheroes.award.AwardRepository;
 
 @RequiredArgsConstructor
 @Service
-public class AwardServiceImpl implements AwardService{
+public class AwardServiceImpl implements AwardService {
 
     private final AwardRepository repository;
     private final AwardMapper mapper;
 
     @Override
+    @Transactional(readOnly = true)
     public Page<AwardDto> getPage(Pageable page) {
         Page<Award> awards = repository.findAll(page);
         return awards.map(mapper::toDto);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public AwardDto getById(Long id) {
         return mapper.toDto(repository.getOne(id));
     }
 
     @Override
+    @Transactional
     public AwardDto create(AwardDto dto) {
         Award award = mapper.toEntity(dto);
         award.setId(null);
@@ -35,6 +39,7 @@ public class AwardServiceImpl implements AwardService{
     }
 
     @Override
+    @Transactional
     public AwardDto update(Long id, AwardDto dto) {
         Award award = repository.getOne(id);
         mapper.toEntityUpdate(dto, award);
@@ -42,7 +47,10 @@ public class AwardServiceImpl implements AwardService{
     }
 
     @Override
+    @Transactional
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        }
     }
 }
